### 2024-01-22 #6257 @editorialbot
[PRE REVIEW]: LabelProp: A semi-automatic segmentation tool for 3D medical images
**Submitting author:** @nathandecaux (<a href="http://orcid.org/0000-0002-6911-6373">Nathan Decaux</a>)
**Repository:** https://github.com/nathandecaux/labelprop
**Branch with paper.md** (empty if default branch):
**Version:** v1.2
