### 2024-01-25 #6257 @WangKehan573
No problem, I can help

---- Replied Message ----
| From | Johanna ***@***.***> |
| Date | 01/25/2024 20:43 |
| To | ***@***.***> |
| Cc | ***@***.***>***@***.***> |
| Subject | Re: [openjournals/joss-reviews] [PRE REVIEW]: LabelProp: A semi-automatic segmentation tool for 3D medical images (Issue #6257) |

@pchlap, @animikhaich, @WangKehan573, could you kindly help us out with reviewing this paper?
