### 2024-01-31 nathandecaux/labelprop#3 @animikhaich
JOSS Review: Missing Dependency - Pandas
* LabelProp version: v1.2
* Python version: 3.11.7
* Operating System: Ubuntu 22.04.3 LTS x86_64

### Description

Missing Pandas dependency.

### What I Did

1. Fresh install in new conda environment.
2. Ran the CLI command: `labelprop pretrain --help`
3. Got the error:
```
Traceback (most recent call last):
  File "/home/ani/miniconda3/envs/joss/bin/labelprop", line 5, in <module>
    from labelprop.cli import cli
  File "/home/ani/miniconda3/envs/joss/lib/python3.11/site-packages/labelprop/cli.py", line 9, in <module>
    import pandas as pd
ModuleNotFoundError: No module named 'pandas'
```

It is a very minor missing dependency but is needed to pass the JOSS Review for self-contained installation. Please add "pandas" as a dependency for the library in the setup.py.

JOSS Review Thread: https://github.com/openjournals/joss-reviews/issues/6284
