### 2024-01-31 nathandecaux/labelprop#4 @animikhaich
JOSS Review: Review Blockers, Updates Requested
* LabelProp version: v1.2
* Python version: 3.11.7
* Operating System: Ubuntu 22.04.3 LTS x86_64

### Description

Required updates to the repository/documentation as a part of the JOSS review process:
1.  **Functionality documentation**: Documentation is currently quite limited and does not explain the complete usage of the library. More details are expected targeting an audience who may not be familiar with labelling medical images. Add some example samples for someone to get started and get an idea of how to use and what to expect.
2. **Community guidelines**: Clear community guidelines and documentation is needed for participants seeking to: 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
3. **Automated tests**: Unit tests are currently missing. This needs to be populated and validated. (https://github.com/nathandecaux/labelprop/blob/master/tests/test_labelprop.py) if there are tests in an alternate file, please share the URL for the same.
4. **State of the field:** The submitted paper draft needs to be updated with a description/section containing how this software compares to other commonly used packages.

JOSS Review Thread: https://github.com/openjournals/joss-reviews/issues/6284
