+++ tests/test_labelprop.py
@@ -5,17 +5,160 @@
 
 import unittest
 
-from labelprop import lightning_model
+from labelprop.napari_entry import train_and_infer
+import numpy as np
+import os
+import nibabel as ni
+import pandas as pd
+import subprocess as sp
+import time
+import json
+import requests
+import io
 
+def create_buf_npz(array_dict):
+    buf = io.BytesIO()
+    np.savez_compressed(buf, **array_dict)
+    buf.seek(0)
+    return buf
+
+def hash_array(array):
+    return hashlib.md5(array.tobytes()).hexdigest()
 
 class TestLabelprop(unittest.TestCase):
     """Tests for `labelprop` package."""
 
     def setUp(self):
-        """Set up test fixtures, if any."""
+        img=np.ones((20,20,20))
+        img[:,0:10,0:10]=0.5
 
-    def tearDown(self):
-        """Tear down test fixtures, if any."""
+        mask=np.zeros((20,20,20))
+        #Add a random square to the mask at the first and last slice
+        mask[0,0:10,0:10]=1
+        mask[-1,0:10,0:10]=1
 
-    def test_000_something(self):
-        """Test something."""
+        hints=np.zeros((20,20,20))
+        hints[:,0:5,0:5]=1
+        #Make dir /tmp/labelprop_test
+        if not os.path.exists('/tmp/labelprop_test'):
+            os.makedirs('/tmp/labelprop_test')
+
+
+        #Save image, mask and hints as nifti
+        ni.save(ni.Nifti1Image(img,np.eye(4)),'/tmp/labelprop_test/img.nii.gz')
+        ni.save(ni.Nifti1Image(mask,np.eye(4)),'/tmp/labelprop_test/mask.nii.gz')
+        ni.save(ni.Nifti1Image(hints,np.eye(4)),'/tmp/labelprop_test/hints.nii.gz')
+
+        #Create a csv for pretraing and training 
+        pretraining_csv=pd.DataFrame({'img_path':['/tmp/labelprop_test/img.nii.gz']})
+        pretraining_csv.to_csv('/tmp/labelprop_test/pretraining.csv',index=False,header=False)
+
+        #Kill any flask server running
+        os.system('pkill -9 flask')
+   
+
+
+    def do_train_and_infer(self):
+        """Test train_and_infer with random image and mask"""
+        img=np.ones((20,20,20))
+        img[:,0:10,0:10]=0.5
+
+        mask=np.zeros((20,20,20))
+        #Add a random square to the mask at the first and last slice
+        mask[0,0:10,0:10]=1
+        mask[-1,0:10,0:10]=1
+        
+        #Add smaller squares in intermediate slices as hints
+        hints=np.zeros((20,20,20))
+        hints[:,0:5,0:5]=1
+        Y_up,Y_down,Y_fused=train_and_infer(img,mask,None,shape=20,max_epochs=1,z_axis=0,output_dir='/tmp/labelprop_test',name='test_function',pretraining=False,hints=hints)
+        return Y_fused
+    def test_train_and_infer(self):
+        """Test train_and_infer with random image and mask"""
+        Y_fused=self.do_train_and_infer()
+        self.assertEqual(Y_fused.shape,(20,20,20))
+        #Assert test.ckpt exists
+        self.assertEqual(os.path.exists('/tmp/labelprop_test/test_function.ckpt'),True)
+
+
+
+    def test_cli(self):
+        """Test pretraining"""
+
+        os.system('labelprop pretrain pretraining.csv --shape 20 --max_epochs 1 --z_axis 0 --output_dir /tmp/labelprop_test --name test_pretraining')
+        #Assert test.ckpt exists
+        self.assertEqual(os.path.exists('/tmp/labelprop_test/test_pretraining.ckpt'),True)
+
[…]
