+++ docs/README.md
@@ -4,9 +4,7 @@
 
 A full list of [Labelprop](https://github.com/nathandecaux/labelprop) project modules.
 
-    - [Dataloading](labelprop/DataLoading.md#dataloading)
-    - [Multilevelnet](labelprop/MultiLevelNet.md#multilevelnet)
-    - [Pretraining Model](labelprop/Pretraining_model.md#pretraining-model)
+
 - [Labelprop](labelprop/index.md#labelprop)
     - [Api](labelprop/api.md#api)
     - [Cli](labelprop/cli.md#cli)
@@ -15,4 +13,5 @@ A full list of [Labelprop](https://github.com/nathandecaux/labelprop) project mo
     - [Train](labelprop/train.md#train)
     - [Utils](labelprop/utils.md#utils)
     - [Voxelmorph2d](labelprop/voxelmorph2d.md#voxelmorph2d)
-    - [Weight Optimizer](labelprop/weight_optimizer.md#weight-optimizer)
+    - [Dataloading](labelprop/DataLoading.md#dataloading)
+    - [Pretraining Model](labelprop/Pretraining_model.md#pretraining-model)
