+++ docs/usage.rst
@@ -1,7 +1,0 @@
-=====
-Usage
-=====
-
-To use LabelProp in a project::
-
-    import labelprop
