+++ .readthedocs.yml
@@ -12,3 +12,8 @@ mkdocs:
 
 # Build all formats
 formats: all
+
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.12"
