+++ README.md
@@ -1,6 +1,6 @@
 # LabelProp : A semi-automatic segmentation tool for 3D medical images
 
-![ReadTheDoc](https://readthedocs.org/projects/labelprop/badge/?version=latest&style=<style>)
+[![ReadTheDoc](https://readthedocs.org/projects/labelprop/badge/?version=latest&style=<style>)](https://labelprop.readthedocs.io/en/latest/)
 [![License](https://img.shields.io/pypi/l/deep-labelprop.svg?color=green)](https://github.com/nathandecaux/labelprop/raw/main/LICENSE)
 [![PyPI](https://img.shields.io/pypi/v/deep-labelprop.svg?color=green)](https://pypi.org/project/deep-labelprop)
 [![Python Version](https://img.shields.io/pypi/pyversions/deep-labelprop.svg?color=green)](https://python.org)
