+++ paper.bib
@@ -1,3 +1,36 @@
+@article{albu2008morphology,
+	title = {A morphology-based approach for interslice interpolation of anatomical slices from volumetric images},
+	volume = {55},
+	doi = {10/fp7xpw},
+	number = {8},
+	journal = {IEEE Transactions on Biomedical Engineering},
+	author = {Albu, Alexandra Branzan and Beugeling, Trevor and Laurendeau, Denis},
+	year = {2008},
+	pages = {2022--2038},
+}
+
+@inproceedings{ronneberger2015u,
+  title={U-net: Convolutional networks for biomedical image segmentation},
+  author={Ronneberger, Olaf and Fischer, Philipp and Brox, Thomas},
+  booktitle={Medical Image Computing and Computer-Assisted Intervention--MICCAI 2015: 18th International Conference, Munich, Germany, October 5-9, 2015, Proceedings, Part III 18},
+  pages={234--241},
+  year={2015},
+  organization={Springer}
+}
+
+@misc{jabri_space-time_2020,
+	title = {Space-{Time} {Correspondence} as a {Contrastive} {Random} {Walk}},
+	url = {http://arxiv.org/abs/2006.14613},
+	doi = {10.48550/arXiv.2006.14613},
+	abstract = {This paper proposes a simple self-supervised approach for learning a representation for visual correspondence from raw video. We cast correspondence as prediction of links in a space-time graph constructed from video. In this graph, the nodes are patches sampled from each frame, and nodes adjacent in time can share a directed edge. We learn a representation in which pairwise similarity defines transition probability of a random walk, so that long-range correspondence is computed as a walk along the graph. We optimize the representation to place high probability along paths of similarity. Targets for learning are formed without supervision, by cycle-consistency: the objective is to maximize the likelihood of returning to the initial node when walking along a graph constructed from a palindrome of frames. Thus, a single path-level constraint implicitly supervises chains of intermediate comparisons. When used as a similarity metric without adaptation, the learned representation outperforms the self-supervised state-of-the-art on label propagation tasks involving objects, semantic parts, and pose. Moreover, we demonstrate that a technique we call edge dropout, as well as self-supervised adaptation at test-time, further improve transfer for object-centric correspondence.},
+	urldate = {2023-09-08},
+	publisher = {arXiv},
+	author = {Jabri, Allan and Owens, Andrew and Efros, Alexei A.},
+	month = dec,
+	year = {2020},
+	note = {arXiv:2006.14613 [cs, eess]},
+	keywords = {Computer Science - Computer Vision and Pattern Recognition, Computer Science - Machine Learning, Electrical Engineering and Systems Science - Image and Video Processing},
+}
 
 @article{decaux_semi-automatic_2023,
 	title = {Semi-automatic muscle segmentation in {MR} images using deep registration-based label propagation},
