+++ README.md
@@ -84,7 +84,7 @@ ni.save(ni.Nifti1Image(fused_propagated_annotations,ni.load('img.nii.gz').affine
 ```
 
 <details open>
-<summary>CLI</summary>
+<summary><h2>CLI</h2></summary>
 Basic operations can be done using the command-line interface provided in labelprop.py at the root of the project.
 
 
@@ -173,7 +173,7 @@ Although Labelprop works on a single scan, it is preferable to pre-train the mod
 
 
 <details open>
-<summary>GUI</summary>
+<summary><h2>GUI</h2></summary>
 See this [repo](https://github.com/nathandecaux/napari-labelprop) to use labelprop main functions in Napari 
 (See also [napari-labelprop-remote](https://github.com/nathandecaux/napari-labelprop-remote) to run labelprop in a separate process, locally or remotely, which uses the [API](https://github.com/nathandecaux/labelprop/blob/master/labelprop/api.py) ).
 
