+++ paper.md
@@ -54,7 +54,7 @@ bibliography: paper.bib
 
 # Summary
 
-LabelProp is a tool that provides a semi-automated method to segment 3D medical images with multiple labels. It is a convenient implementation of our peer-reviewed method designed to assist medical professionals in segmenting musculoskeletal structures on scans based on a small number of annotated slices [@decaux_semi-automatic_2023]. LabelProp leverages deep learning techniques, but does not require a training set or pretrained weights to function. It is available as a PyPi package and offers both a command-line interface (CLI) and an API. Additionally, LabelProp provides two plugins, namely 'napari-labelprop' and 'napari-labelprop-remote', which facilitate training and inference on a single scan within the multi-dimensional viewer Napari. It is available on GitHub with pretrained weights ([https://github.com/nathandecaux/napari-labelprop](https://github.com/nathandecaux/napari-labelprop))
+LabelProp is a tool that provides a semi-automated method to segment 3D medical images with multiple labels. It is a convenient implementation of our peer-reviewed method designed to assist medical professionals in segmenting musculoskeletal structures on scans based on a small number of annotated slices [@decaux_semi-automatic_2023]. LabelProp leverages deep learning techniques, but can be used without a training set. It is available as a PyPi package and offers both a command-line interface (CLI) and an API. Additionally, LabelProp provides two plugins, namely 'napari-labelprop' and 'napari-labelprop-remote', which facilitate training and inference on a single scan within the multi-dimensional viewer Napari. It is available on GitHub with pretrained weights ([https://github.com/nathandecaux/napari-labelprop](https://github.com/nathandecaux/napari-labelprop))
 
 # Statement of need
 
@@ -64,19 +64,19 @@ LabelProp addresses these challenges with a novel deep registration-based label 
 
 A key innovation of LabelProp is its minimal reliance on manual annotations. Demonstrating the capability for accurate 3D segmentation from as few as three annotated slices per MR volume [@decaux_semi-automatic_2023], it significantly reduces the workload for medical professionals and is particularly beneficial where extensive annotated data is scarce. This feature aligns with the method of slice-to-slice registration [@ogier2017individual], but is further enhanced by deep learning techniques.
 
-Similar to Voxelmorph, the underlying model in this approach learns to generate deformations without supervision [@balakrishnan2019voxelmorph]. However, it specifically focuses on aligning adjacent 2D slices and can be trained directly on the scan that needs to be segmented or on a complete dataset for optimal results. When training the model with at least two annotations for a scan, a constraint is added to ensure that the generated deformations are consistent from both an image and segmentation perspective. Additionally, weak annotations in the form of scribbles can be provided during training on intermediate slices to provide additional guidance for propagation. Examples of manual annotations and "scribbles" are shown in Fig. 1. 
+Similar to Voxelmorph, the underlying model in this approach learns to generate deformations without supervision [@balakrishnan2019voxelmorph]. However, it specifically focuses on aligning adjacent 2D slices and can be trained directly on the scan that needs to be segmented or on a complete dataset for optimal results. When training the model with at least two annotations for a scan, a constraint is added to ensure that the generated deformations are consistent from both an image and segmentation perspective. Additionally, weak annotations in the form of scribbles can be provided during training on intermediate slices to provide additional guidance for propagation. Examples of manual annotations and scribbles are shown in Fig. 1. 
 
[…]
