### 2024-02-12 nathandecaux/labelprop#4 @nathandecaux
Thanks for your feedback, which we have taken into account :

1.  The "Data" and "Basic Use" sub-sections have been added to the main README to better describe the required data and the expected results while using Labelprop main features.
2. A "How to contribute" section has been added to the main README. In addition, a read-the-doc website is now provided [here](https://labelprop.readthedocs.io/en/latest/)
3. [test_labelprop.py](https://github.com/nathandecaux/labelprop/blob/master/tests/test_labelprop.py) has been populated with several unit tests.
4. A "State of the field" section has been added to the paper.
