+++ setup.py
@@ -54,6 +54,6 @@ setup(
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/nathandecaux/labelprop',
-    version='1.2.1.1',
+    version='1.2.2',
     zip_safe=False,
 )
