### 2024-02-12 #6284 @nathandecaux
Hello.The requested updates have been taken into account in the [v1.2.2 release](https://github.com/nathandecaux/labelprop/releases/tag/v1.2.2).
