### 2024-04-04 eec491b
Optimizing resampling method
