### 2024-04-11 2fbe5c9
Optimizing few stuff
