### 2024-09-23 #6284 @likeajumprope
> Hi @likeajumprope I assume no joy over email?

Not yet, but will follow up again and send out some more
