### 2024-09-30 #6284 @rousseau
Dear editor,
I understand that the peer review process can take time, but I would like to inquire about the status of our article's evaluation.
Please let me know if you require any further information from us.
Thank you for your time and consideration.
Sincerely.
