### 2024-10-27 #6284 @animikhaich
Hi @samhforbes, One of my connections is happy to take the place of the second reviewer. Please proceed with the necessary formalities to onboard him. His profile is: https://github.com/tensorsofthewall

Some highlights about him:
- We collaborated on a recent project https://github.com/tensorsofthewall/VidTune
- He has recently authored a paper at ECCV 2024. Details: https://unilcd.github.io/

Please let me know if anything else is needed from my end. Thanks!
