### 2024-10-27 #6284 @animikhaich
Hi @likeajumprope, Sure:
- We have never published together.
- We have worked in the same research lab, but on different projects/teams.
- We had common courses during our MS degree.

Let me know if you have any other questions. Thanks!
