### 2025-01-16 #6284 @nathandecaux
> # Post-Review Checklist for Editor and Authors
> ## Additional Author Tasks After Review is Complete
> * Double check authors and affiliations (including ORCIDs)
> * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> * Archive the release on Zenodo/figshare/etc and post the DOI here.
> * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * Make sure that the license listed for the archive is the same as the software license.
>
> ## Editor Tasks Prior to Acceptance
> * [ ]  Read the text of the paper and offer comments/corrections (as either a list or a pull request)[ ]  Check that the archive title, author list, version tag, and the license are correct[ ]  Set archive DOI with `@editorialbot set <DOI here> as archive`[ ]  Set version with `@editorialbot set <version here> as version`[ ]  Double check rendering of paper with `@editorialbot generate pdf`[ ]  Specifically check the references with `@editorialbot check references` and ask author(s) to update as needed[ ]  Recommend acceptance with `@editorialbot recommend-accept`

Dear editor,
Thank you for the checklist and guidance. I have completed the required tasks following the review process:

1. **Authors and Affiliations**: I have double-checked the authors, affiliations, and ORCIDs, and they are correct.
2. **Software Release**: The post-review version of the software is **v1.2.4**, which includes all changes from the review process. The release is available [here](https://github.com/nathandecaux/labelprop/releases/tag/v1.2.4).
3. **Archiving on Zenodo**: The software has been archived on Zenodo. The DOI for the archive is **[10.5281/zenodo.14674566](https://doi.org/10.5281/zenodo.14674566)**.
4. **Consistency Checks**: I have ensured that:
   - The title and author list (including ORCIDs) in the Zenodo archive match those in the JOSS paper.
   - The license listed in the archive matches the software license. (CC-BY-4.0)

Please let me know if there is anything else required on my end.

Best regards,
Nathan Decaux
