### 2025-02-07 #6284 @likeajumprope
> Dear editor,
>
> [@nathandecaux](https://github.com/nathandecaux) has completed the required tasks.
>
> Is there anything else to do before final publication ?
>
> Best regards, François

Thanks for the ping @rousseau  - this one is on my list for the weekend :)
