### 2025-02-09 #6284 @likeajumprope
@rousseau @nathandecaux could you:
- [ ] change Voxelmorph in line 35 to VoxelMorph
- [ ] provide a link for the MyoSegmenTUM data set
- [ ] add DOIS for the Ronneberger, Ogier and Sakinis citations

We're close!
