### 2025-02-13 #6284 @likeajumprope
> > [@rousseau](https://github.com/rousseau) [@nathandecaux](https://github.com/nathandecaux) could you:
> > ```
> > * [ ]  change Voxelmorph in line 35 to VoxelMorph[ ]  provide a link for the MyoSegmenTUM data set[ ]  add DOIS for the Ronneberger, Ogier and Sakinis citations
> > ```
> >
> >
> >
> >
> >
> >
> >
> >
> >
> >
> >
> > We're close!
>
> Dear editor, The requested changes have been taken into account in the latest push of the main branch. Do I need to create a new release as well? Regards, Nathan Decaux

Yes please
