### 2025-02-13 #6284 @nathandecaux
I

> > > [@rousseau](https://github.com/rousseau) [@nathandecaux](https://github.com/nathandecaux) could you:
> > > ```
> > > * [ ]  change Voxelmorph in line 35 to VoxelMorph[ ]  provide a link for the MyoSegmenTUM data set[ ]  add DOIS for the Ronneberger, Ogier and Sakinis citations
> > > ```
> > >
> > >
> > >
> > >
> > >
> > >
> > >
> > >
> > >
> > >
> > >
> > > We're close!
> >
> >
> > Dear editor, The requested changes have been taken into account in the latest push of the main branch. Do I need to create a new release as well? Regards, Nathan Decaux
>
> Yes please

Ok, I've created the release with the tag [v1.2.4-joss](https://github.com/nathandecaux/labelprop/releases/tag/v1.2.4-joss)
Regards,
Nathan Decaux
