### 2025-05-08 #6284 @danielskatz
> [@editorialbot](https://github.com/editorialbot) set zenodo.14674566 as archive

This isn't the right name. The prefix is missing.  It should be `@editorialbot set 10.5281/zenodo.14674566 as archive`
