### 2025-05-08 #6284 @likeajumprope
> **License info**:
>
> 🔴 License found: `Creative Commons Attribution 4.0 International` (Not [OSI approved](https://opensource.org/licenses))

will check if the license is ok
