### 2025-05-19 #6284 @likeajumprope
> Dear editor [@likeajumprope](https://github.com/likeajumprope),
>
> Do we need to change anything in the article?
>
> Best,

No, the license should be fine
