### 2025-05-29 #6284 @samhforbes
@nathandecaux @danielskatz flagged the issue above:

> There are a bunch of problems in the bib file that need to be fixed, as shown by the errors/missing DOI in the above two comments from editorialbot.

For the last set of errors, these are not DOIs but handles, so the DOI field should changed to the actual DOI. For example, the DOI corresponding to the first one is 10.1109/TBME.2008.921158

I also see other stuff in the bib file that should be cleaned up, such as the forced casing in this
	title = {An {Analysis} of the {Impact} of {Annotation} {Errors} on the {Accuracy} of {Deep} {Learning} for {Cell} {Segmentation}},
Forced casing should only be used when JOSS process will make something lowercase that shouldn't be, not to override the JOSS style in general, which is Sentence case for article titles.

If you fix these it should compile.
