### 2025-05-30 #6284 @nathandecaux
Dear editor,
The release [v1.2.4-joss ](https://github.com/nathandecaux/labelprop/releases/tag/v1.2.4-joss) has been recreated with the corrected bibliography.
Regards,
Nathan Decaux
