### 2025-05-30 #6284 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
