### 2025-05-30 #6284 @samhforbes
@nathandecaux only thing to change from what I can see is that errant DOI, which is Chanti et al, and should be I think 10.1109/TMI.2021.3058303
